+++ checker-qual/src/main/java/org/checkerframework/common/util/count/report/qual/ReportOverride.java
@@ -1,4 +1,4 @@
-package org.checkerframework.common.util.report.qual;
+package org.checkerframework.common.util.count.report.qual;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -6,8 +6,8 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-/** Report all write accesses to a field with this annotation. */
+/** Report all methods that override a method with this annotation. */
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.FIELD)
-public @interface ReportWrite {}
+@Target(ElementType.METHOD)
+public @interface ReportOverride {}