+++ checker-qual/src/main/java/org/checkerframework/common/util/count/report/qual/ReportUse.java
@@ -1,4 +1,4 @@
-package org.checkerframework.common.util.report.qual;
+package org.checkerframework.common.util.count.report.qual;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -6,8 +6,8 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-/** Report all methods that override a method with this annotation. */
+/** Report all uses of a type that has this annotation. Can also be used on a package. */
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface ReportOverride {}
+@Target({ElementType.PACKAGE, ElementType.TYPE})
+public @interface ReportUse {}