+++ checker-qual/src/main/java/org/checkerframework/common/util/count/report/qual/ReportWrite.java
@@ -1,4 +1,4 @@
-package org.checkerframework.common.util.report.qual;
+package org.checkerframework.common.util.count.report.qual;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -6,8 +6,8 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-/** Report all uses of a type that has this annotation. Can also be used on a package. */
+/** Report all write accesses to a field with this annotation. */
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.PACKAGE, ElementType.TYPE})
-public @interface ReportUse {}
+@Target(ElementType.FIELD)
+public @interface ReportWrite {}