+++ checker-qual/src/main/java/org/checkerframework/common/util/count/report/qual/ReportInherit.java
@@ -1,4 +1,4 @@
-package org.checkerframework.common.util.report.qual;
+package org.checkerframework.common.util.count.report.qual;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -6,8 +6,8 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-/** Report all read or write access to a field with this annotation. */
+/** Report all types that extend/implement a type that has this annotation. */
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.FIELD)
-public @interface ReportReadWrite {}
+@Target(ElementType.TYPE)
+public @interface ReportInherit {}