+++ framework/src/main/java/org/checkerframework/common/util/debug/DoNothingProcessor.java
@@ -8,16 +8,19 @@ import javax.lang.model.SourceVersion;
 import javax.lang.model.element.TypeElement;
 
 /**
- * Empty simple processor.
+ * This is an annotation processor that does nothing.
  *
- * <p>It is useful in debugging compiler behavior with an annotation processor present.
+ * <p>Use it when you are required to provide an annotation processor, or when you want to debug
+ * compiler behavior with an annotation processor present.
  */
 @SupportedAnnotationTypes("*")
-public class EmptyProcessor extends AbstractProcessor {
+public class DoNothingProcessor extends AbstractProcessor {
+
+  /** Creates a DoNothingProcessor. */
+  public DoNothingProcessor() {}
 
   @Override
   public boolean process(Set<? extends TypeElement> annotations, RoundEnvironment roundEnv) {
-    System.out.println("Empty Processor run!");
     return false;
   }