+++ docs/CHANGELOG.md
@@ -1,3 +1,13 @@
+Version 3.43.0 (January 2, 2024)
+--------------------------------
+
+**User-visible changes:**
+
+**Implementation details:**
+
+**Closed issues:**
+
+
 Version 3.42.0 (December 15, 2023)
 ----------------------------------